Finishes constructing the InventoryItem class
This commit changes how the system handles inventory items. It introduces
a new throwable, InventoryStackFullError. The InventoryItem class is modified
to accept a max_stack parameter in it's constructor. When attempting to
incrementCount, if the statck is full it throws said error. In the hero class,
this throwable is accounted for. Introduces Issue # 5