Refactors how changing room logic works.
The Character.setLocation() method is now Character.move(). This method is overriden in
the Hero class to automatically set the last location and handle that. Therefore, all
normal moves of the player should be just a single call to Hero.move(). If the hero has
to return to the last room, the Hero.bounceBack() method should be called. All occurances
of any of the aforementioned methods have already been refactored and adjusted.